Extract Ranking (#572)
