+++ vj4/app.py
@@ -94,6 +94,7 @@ class Application(web.Application):
     from vj4.handler import problem
     from vj4.handler import record
     from vj4.handler import training
+    from vj4.handler import ranking
     from vj4.handler import user
     from vj4.handler import i18n
     if options.static:
