+++ vj4/handler/user.py
@@ -254,20 +254,3 @@ class UserSearchHandler(base.Handler):
     for i in range(len(udocs)):
       self.modify_udoc(udocs, i)
     self.json(udocs)
-
-
-@app.route('/ranking', 'domain_ranking')
-class RankingHandler(base.Handler):
-  USERS_PER_PAGE = 100
-
-  @base.require_priv(builtin.PRIV_USER_PROFILE)
-  @base.require_perm(builtin.PERM_VIEW_RANKING)
-  @base.get_argument
-  @base.route_argument
-  @base.sanitize
-  async def get(self, *, page: int=1):
-    dudocs, dupcount, _ = await pagination.paginate(
-        domain.get_multi_user(domain_id=self.domain_id, rp={'$gt': 0.0}).sort([('rank', 1)]),
-        page, self.USERS_PER_PAGE)
-    udict = await user.get_dict(dudoc['uid'] for dudoc in dudocs)
-    self.render('domain_ranking.html', page=page, dupcount=dupcount, dudocs=dudocs, udict=udict)
