+++ vj4/ui/pages/ranking_main.page.styl
@@ -1,4 +1,4 @@
-.page--domain_ranking
+.page--ranking_main
   .col--rank
     width: rem(90px)
     border-right: 1px solid $table-border-color
