+++ vj4/ui/templates/partials/nav.html
@@ -29,7 +29,7 @@
       {{ nav_item('homework_main', 'homework') }}
     {% endif %}
     {% if handler.has_priv(vj4.model.builtin.PRIV_USER_PROFILE) and handler.has_perm(vj4.model.builtin.PERM_VIEW_RANKING) %}
-      {{ nav_item('domain_ranking', 'domain_ranking') }}
+      {{ nav_item('ranking_main', 'ranking_main') }}
     {% endif %}
     {% if handler.has_perm(vj4.model.builtin.PERM_EDIT_DESCRIPTION) or handler.has_perm(vj4.model.builtin.PERM_EDIT_PERM) %}
       {{ nav_item('domain_manage', 'domain_manage') }}
